[39334651] Checkouts using PagSeguro
In this story, PagSeguro was implemented and payments are now being
done correctly.

Some aspects of the sign up form had to be changed. PagSeguro
requires that address number and phone code area be given separated, so some
changes had to be made rather sooner than later.

The notifications from PagSeguro are still missing. I'm going to do
that in the next commit.