[38314589] Fixes user registration handle sanitization
We changed the strategy when sanitizing the handle in the user
registration form. Before this, the input had its value changed on the
fly, which means arrow keys wouldn't work.

We know filter the keyCode of each event, allowing only the values that
we want (letters, number, backspace, delete, arrow keys etc).