[29128519] Adds creation and edition of taxonomies
This commit adds the creation and edition features in the admin panel.
It's very rought at this point, and the Javascript code is really ugly
and lacking better tests. However, given we should move forward as soon
as possible and this is working right now, it's ok.

When this starts to hurt, then we can write some tests and refactor.