[42446519] Creates the offline sales page
This commits uses Ember.js 1.0.0pre4 for the offline sales page. Given
it was mostly an experience with Ember, we're creating orders right
away. Next, we'll create carts instead, which will then be converted
into orders.

The api in `/admin/api/v1` was also created to deal with requests from
the offline sales page.

Given we added this offline sales page, we also added a new field called
environments into the `Order` model. This can contain, for now, either
`website` or `offline`.