[47858093] Allows Customers to be created at the POS
This commit adds the ability to create a customer at the POS or anywhere
(via the API). In these kind of situations, the user can't enter all
his information, so only first name, last name and social security
number are required.

Email is not required either. For now, if the user want to login to the website
later, he'll have to go to the physical store to add an email to its
account. The ergonomics of that process can be improved later.